Add files via upload
Date: November 8th 2018
Location: Apartment in Denton
Description: A picture of some text on a screen. A definition is found within the text.
File origin: My phone but the text is owned by the publisher of that article
Intended Purpose: To help clarify a concept for me and some members of a group in hopes that we have an easy reference.